feat: explode vu1 in files feat: added way  more tests for vu1
