+++ ps2xRuntime/CMakeLists.txt
@@ -358,7 +358,9 @@ add_library(ps2_runtime STATIC
     src/lib/ps2_pad.cpp
     src/lib/ps2_runtime.cpp
     src/lib/ps2_vif1_interpreter.cpp
-    src/lib/ps2_vu1.cpp
+    src/lib/vu/ps2_vu1_core.cpp
+    src/lib/vu/ps2_vu1_upper.cpp
+    src/lib/vu/ps2_vu1_lower.cpp
     src/lib/games_database.cpp
 )
