+++ ps2xRuntime/src/lib/vu/ps2_vu1_upper.cpp
@@ -1,0 +1,458 @@
+#include "runtime/ps2_vu1.h"
+#include "ps2_vu1_detail.h"
+
+#include <cmath>
+#include <cstring>
+
+// ============================================================================
+// Upper instructions (FMAC pipeline)
+// ============================================================================
+void VU1Interpreter::execUpper(uint32_t instr)
+{
+    uint8_t dest = DEST(instr);
+    uint8_t ft = FT(instr);
+    uint8_t fs = FS(instr);
+    uint8_t fd = FD(instr);
+    uint8_t op = instr & 0x3F;
+
+    float *vd = m_state.vf[fd];
+    const float *vs = m_state.vf[fs];
+    const float *vt = m_state.vf[ft];
+    float result[4];
+
+    // Upper opcode decoding (bits 5:0 of upper word)
+    switch (op)
+    {
+    case 0x00:
+    case 0x01:
+    case 0x02:
+    case 0x03: // ADDbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] + bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x04:
+    case 0x05:
+    case 0x06:
+    case 0x07: // SUBbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] - bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x08:
+    case 0x09:
+    case 0x0A:
+    case 0x0B: // MADDbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] + vs[c] * bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x0C:
+    case 0x0D:
+    case 0x0E:
+    case 0x0F: // MSUBbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] - vs[c] * bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x10:
+    case 0x11:
+    case 0x12:
+    case 0x13: // MAXbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] > bc) ? vs[c] : bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x14:
+    case 0x15:
+    case 0x16:
+    case 0x17: // MINIbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] < bc) ? vs[c] : bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x18:
+    case 0x19:
+    case 0x1A:
+    case 0x1B: // MULbc
+    {
+        float bc = broadcast(vt, op & 3);
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] * bc;
+        applyDest(vd, result, dest);
+        return;
+    }
+    case 0x1C: // MULq
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] * m_state.q;
+        applyDest(vd, result, dest);
+        return;
+    case 0x1D: // MAXi
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] > m_state.i) ? vs[c] : m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x1E: // MULi
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] * m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x1F: // MINIi
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] < m_state.i) ? vs[c] : m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x20: // ADDq
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] + m_state.q;
+        applyDest(vd, result, dest);
+        return;
+    case 0x21: // MADDq
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] + vs[c] * m_state.q;
+        applyDest(vd, result, dest);
+        return;
+    case 0x22: // ADDi
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] + m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x23: // MADDi
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] + vs[c] * m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x24: // SUBq
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] - m_state.q;
+        applyDest(vd, result, dest);
+        return;
+    case 0x25: // MSUBq
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] - vs[c] * m_state.q;
+        applyDest(vd, result, dest);
+        return;
+    case 0x26: // SUBi
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] - m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x27: // MSUBi
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] - vs[c] * m_state.i;
+        applyDest(vd, result, dest);
+        return;
+    case 0x28: // ADD
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] + vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x29: // MADD
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] + vs[c] * vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x2A: // MUL
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] * vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x2B: // MAX
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] > vt[c]) ? vs[c] : vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x2C: // SUB
+        for (int c = 0; c < 4; c++)
+            result[c] = vs[c] - vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x2D: // MSUB
+        for (int c = 0; c < 4; c++)
+            result[c] = m_state.acc[c] - vs[c] * vt[c];
+        applyDest(vd, result, dest);
+        return;
+    case 0x2E: // OPMSUB
+        result[0] = m_state.acc[0] - vs[1] * vt[2];
+        result[1] = m_state.acc[1] - vs[2] * vt[0];
+        result[2] = m_state.acc[2] - vs[0] * vt[1];
+        result[3] = 0.0f;
+        applyDest(vd, result, dest);
+        return;
+    case 0x2F: // MINI
+        for (int c = 0; c < 4; c++)
+            result[c] = (vs[c] < vt[c]) ? vs[c] : vt[c];
+        applyDest(vd, result, dest);
+        return;
+
+    // Upper special group (low op 0x3C..0x3F).
+    // Like lower1 special, the real selector is not just bits 5:0.  Dobie decodes:
+    //   op = (instr & 0x3) | ((instr >> 4) & 0x7C)
+    // Several instructions in this group also use FT as the destination, not FD.
+    case 0x3C:
+    case 0x3D:
+    case 0x3E:
+    case 0x3F:
+    {
+        const uint8_t specialOp = static_cast<uint8_t>((instr & 0x3u) | ((instr >> 4) & 0x7Cu));
+        float *vtDest = m_state.vf[ft];
+
+        switch (specialOp)
+        {
+        case 0x00:
+        case 0x01:
+        case 0x02:
+        case 0x03: // ADDAbc
+        {
+            float bc = broadcast(vt, specialOp & 3);
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] + bc;
+            applyDestAcc(result, dest);
+            return;
+        }
+        case 0x04:
+        case 0x05:
+        case 0x06:
+        case 0x07: // SUBAbc
+        {
+            float bc = broadcast(vt, specialOp & 3);
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] - bc;
+            applyDestAcc(result, dest);
+            return;
+        }
+        case 0x08:
+        case 0x09:
+        case 0x0A:
+        case 0x0B: // MADDAbc
+        {
+            float bc = broadcast(vt, specialOp & 3);
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] + vs[c] * bc;
+            applyDestAcc(result, dest);
+            return;
+        }
+        case 0x0C:
+        case 0x0D:
+        case 0x0E:
+        case 0x0F: // MSUBAbc
+        {
+            float bc = broadcast(vt, specialOp & 3);
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] - vs[c] * bc;
+            applyDestAcc(result, dest);
+            return;
+        }
+        case 0x10: // ITOF0
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv;
+                std::memcpy(&iv, &vs[c], 4);
+                result[c] = static_cast<float>(iv);
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x11: // ITOF4
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv;
+                std::memcpy(&iv, &vs[c], 4);
+                result[c] = static_cast<float>(iv) / 16.0f;
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x12: // ITOF12
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv;
+                std::memcpy(&iv, &vs[c], 4);
+                result[c] = static_cast<float>(iv) / 4096.0f;
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x13: // ITOF15
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv;
+                std::memcpy(&iv, &vs[c], 4);
+                result[c] = static_cast<float>(iv) / 32768.0f;
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x14: // FTOI0
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv = static_cast<int32_t>(vs[c]);
+                std::memcpy(&result[c], &iv, 4);
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x15: // FTOI4
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv = static_cast<int32_t>(vs[c] * 16.0f);
+                std::memcpy(&result[c], &iv, 4);
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x16: // FTOI12
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv = static_cast<int32_t>(vs[c] * 4096.0f);
+                std::memcpy(&result[c], &iv, 4);
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x17: // FTOI15
+            for (int c = 0; c < 4; c++)
+            {
+                int32_t iv = static_cast<int32_t>(vs[c] * 32768.0f);
+                std::memcpy(&result[c], &iv, 4);
+            }
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x18:
+        case 0x19:
+        case 0x1A:
+        case 0x1B: // MULAbc
+        {
+            float bc = broadcast(vt, specialOp & 3);
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] * bc;
+            applyDestAcc(result, dest);
+            return;
+        }
+        case 0x1C: // MULAq
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] * m_state.q;
+            applyDestAcc(result, dest);
+            return;
+        case 0x1D: // ABS
+            for (int c = 0; c < 4; c++)
+                result[c] = std::fabs(vs[c]);
+            applyDest(vtDest, result, dest);
+            return;
+        case 0x1E: // MULAi
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] * m_state.i;
+            applyDestAcc(result, dest);
+            return;
+        case 0x1F: // CLIP
+        {
+            float w = std::fabs(vt[3]);
+            uint32_t flags = 0;
+            if (vs[0] > +w) flags |= 0x01;
+            if (vs[0] < -w) flags |= 0x02;
+            if (vs[1] > +w) flags |= 0x04;
+            if (vs[1] < -w) flags |= 0x08;
+            if (vs[2] > +w) flags |= 0x10;
+            if (vs[2] < -w) flags |= 0x20;
+            m_state.clip = (m_state.clip << 6) | flags;
+            return;
+        }
+        case 0x20: // ADDAq
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] + m_state.q;
+            applyDestAcc(result, dest);
+            return;
+        case 0x21: // MADDAq
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] + vs[c] * m_state.q;
+            applyDestAcc(result, dest);
+            return;
+        case 0x22: // ADDAi
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] + m_state.i;
+            applyDestAcc(result, dest);
+            return;
+        case 0x23: // MADDAi
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] + vs[c] * m_state.i;
+            applyDestAcc(result, dest);
+            return;
+        case 0x24: // SUBAq
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] - m_state.q;
+            applyDestAcc(result, dest);
+            return;
+        case 0x25: // MSUBAq
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] - vs[c] * m_state.q;
+            applyDestAcc(result, dest);
+            return;
+        case 0x26: // SUBAi
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] - m_state.i;
+            applyDestAcc(result, dest);
+            return;
+        case 0x27: // MSUBAi
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] - vs[c] * m_state.i;
+            applyDestAcc(result, dest);
+            return;
+        case 0x28: // ADDA
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] + vt[c];
+            applyDestAcc(result, dest);
+            return;
+        case 0x29: // MADDA
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] + vs[c] * vt[c];
+            applyDestAcc(result, dest);
+            return;
+        case 0x2A: // MULA
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] * vt[c];
+            applyDestAcc(result, dest);
+            return;
+        case 0x2C: // SUBA
+            for (int c = 0; c < 4; c++)
+                result[c] = vs[c] - vt[c];
+            applyDestAcc(result, dest);
+            return;
+        case 0x2D: // MSUBA
+            for (int c = 0; c < 4; c++)
+                result[c] = m_state.acc[c] - vs[c] * vt[c];
+            applyDestAcc(result, dest);
+            return;
+        case 0x2E: // OPMULA
+            result[0] = vs[1] * vt[2];
+            result[1] = vs[2] * vt[0];
+            result[2] = vs[0] * vt[1];
+            result[3] = 0.0f;
+            applyDestAcc(result, dest);
+            return;
+        case 0x2F:
+        case 0x30: // NOP
+            return;
+        default:
+            return;
+        }
+    }
+
+    case 0x30:
+    case 0x31:
+    case 0x32:
+    case 0x33:
+    default:
+        return;
+    }
+}
