+++ ps2xTest/CMakeLists.txt
@@ -23,6 +23,7 @@ add_library(ps2_test_lib STATIC
     src/ps2_runtime_kernel_tests.cpp
     src/ps2_runtime_interrupt_tests.cpp
     src/ps2_memory_tests.cpp
+    src/ps2_vu1_tests.cpp
     src/ps2_gs_tests.cpp
     src/ps2_sif_rpc_tests.cpp
     src/ps2_sif_dma_tests.cpp
