+++ ps2xTest/src/main.cpp
@@ -10,6 +10,7 @@ void register_ps2_runtime_io_tests();
 void register_ps2_runtime_kernel_tests();
 void register_ps2_runtime_interrupt_tests();
 void register_ps2_memory_tests();
+void register_ps2_vu1_tests();
 void register_ps2_gs_tests();
 void register_ps2_sif_rpc_tests();
 void register_ps2_sif_dma_tests();
@@ -29,6 +30,7 @@ int main()
     register_ps2_runtime_kernel_tests();
     register_ps2_runtime_interrupt_tests();
     register_ps2_memory_tests();
+    register_ps2_vu1_tests();
     register_ps2_gs_tests();
     register_ps2_sif_rpc_tests();
     register_ps2_sif_dma_tests();
